backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
pyarrow
scikit-learn
scipy
joblib
numpy
python-dotenv
